requirements.txt
Flask>=1.1.2
idna
isort
requests
urllib3
mysql-connector
